refactor(tests): split test_django_db.py into a focused django_db package (#443)
tests/test_django_db.py was 913 LOC, over the 500 module-health
threshold. Split it verbatim into tests/django_db/ — one file per
concern, no production change and no assertion change.

- conftest.py: the module-level autouse _isolate_bad_artifacts fixture,
  lifted unchanged so it keeps applying to every split file
- _shared.py: _make_cfg / _make_importer / _ok_run / _fail_run, imported
  rather than duplicated
- test_helpers.py: pure / DSLR / Postgres helpers + dump validation
- test_migration.py: reference-DB migration with selective faking
- test_restore_pipeline.py: restore-and-copy + DSLR restore
- test_strategies.py: local / remote / CI dump strategies
- test_orchestration.py: full django_db_import orchestration
- test_pruning.py: DSLR snapshot parsing and pruning

Collected test count unchanged (81), Class::method multiset and AST
defs identical. Relates-to #443
