+++ tests/django_db/test_migration.py
@@ -1,0 +1,103 @@
+"""Tests for teatree.utils.django_db — reference-DB migration with selective faking.
+
+Split verbatim from the former monolithic ``tests/test_django_db.py``
+(souliane/teatree#443). No behavior change.
+"""
+
+from pathlib import Path
+from subprocess import CompletedProcess
+
+import pytest
+
+from teatree.utils import run as run_mod
+from teatree.utils.django_db import _MigrateResult
+
+from ._shared import _make_importer, _ok_run
+
+
+class TestMigrateReferenceDb:
+    def test_succeeds_on_first_try(self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+        (tmp_path / "manage.py").write_text("", encoding="utf-8")
+        monkeypatch.setattr(run_mod.subprocess, "run", _ok_run)
+        assert _make_importer(tmp_path)._migrate_reference_db() is _MigrateResult.APPLIED
+
+    def test_returns_already_migrated_when_no_migrations(self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+        (tmp_path / "manage.py").write_text("", encoding="utf-8")
+        monkeypatch.setattr(
+            run_mod.subprocess,
+            "run",
+            lambda *a, **kw: CompletedProcess(a, 0, "No migrations to apply.\n", ""),
+        )
+        assert _make_importer(tmp_path)._migrate_reference_db() is _MigrateResult.ALREADY_MIGRATED
+
+    def test_fakes_already_exists(self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+        (tmp_path / "manage.py").write_text("", encoding="utf-8")
+        calls: list[list[str]] = []
+        call_count = 0
+
+        def fake_run(args, **kw):
+            nonlocal call_count
+            calls.append(list(args))
+            call_count += 1
+            if call_count == 1:
+                return CompletedProcess(args, 1, "Applying myapp.0005_add_field...\n", "already exists")
+            return CompletedProcess(args, 0, "", "")
+
+        monkeypatch.setattr(run_mod.subprocess, "run", fake_run)
+        assert _make_importer(tmp_path)._migrate_reference_db() is _MigrateResult.APPLIED
+        assert "--fake" in calls[1]
+
+    def test_skips_on_config_error(self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+        (tmp_path / "manage.py").write_text("", encoding="utf-8")
+        monkeypatch.setattr(
+            run_mod.subprocess,
+            "run",
+            lambda *a, **kw: CompletedProcess(a, 1, "", "ModuleNotFoundError: No module named 'foo'"),
+        )
+        assert _make_importer(tmp_path)._migrate_reference_db() is _MigrateResult.FAILED
+
+    def test_skips_on_non_fakeable_error(self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+        (tmp_path / "manage.py").write_text("", encoding="utf-8")
+        monkeypatch.setattr(
+            run_mod.subprocess,
+            "run",
+            lambda *a, **kw: CompletedProcess(a, 1, "", "unexpected error"),
+        )
+        assert _make_importer(tmp_path)._migrate_reference_db() is _MigrateResult.FAILED
+
+    def test_skips_when_no_manage_py(self, tmp_path: Path) -> None:
+        assert _make_importer(tmp_path)._migrate_reference_db() is _MigrateResult.ALREADY_MIGRATED
+
+    def test_skips_when_failing_migration_not_parseable(self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+        (tmp_path / "manage.py").write_text("", encoding="utf-8")
+        monkeypatch.setattr(
+            run_mod.subprocess,
+            "run",
+            lambda *a, **kw: CompletedProcess(a, 1, "Running migrate...\n", "already exists"),
+        )
+        assert _make_importer(tmp_path)._migrate_reference_db() is _MigrateResult.FAILED
+
+    def test_exhausts_retries(self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+        (tmp_path / "manage.py").write_text("", encoding="utf-8")
+        monkeypatch.setattr(
+            run_mod.subprocess,
+            "run",
+            lambda *a, **kw: CompletedProcess(a, 1, "Applying myapp.0001_init...\n", "already exists"),
+        )
+        assert _make_importer(tmp_path)._migrate_reference_db() is _MigrateResult.FAILED
+
+    def test_fakes_does_not_exist(self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+        (tmp_path / "manage.py").write_text("", encoding="utf-8")
+        calls: list[list[str]] = []
+        call_count = 0
+
+        def fake_run(args, **kw):
+            nonlocal call_count
+            calls.append(list(args))
+            call_count += 1
+            if call_count == 1:
+                return CompletedProcess(args, 1, "Applying myapp.0005_drop...\n", "does not exist")
+            return CompletedProcess(args, 0, "", "")
+
+        monkeypatch.setattr(run_mod.subprocess, "run", fake_run)
+        assert _make_importer(tmp_path)._migrate_reference_db() is _MigrateResult.APPLIED
